crud jpa pronto

+++ HackStudio/src/java/br/com/hackstudio/dao/Dao.java
@@ -5,7 +5,7 @@ import java.util.List;
 public interface Dao {   
     public boolean save(Object object);
     public boolean update(Object object);
-    public boolean delete(int id);
-    public Object get(int id);
+    public boolean delete(Long id);
+    public Object get(Long id);
     public List<Object> get();    
 }
